Fixed adding of line of controllers when control ruler is zoomed.
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@4813 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ gui/controlruler.h
@@ -125,7 +125,7 @@ protected:
 
     int m_maxItemValue;
 
-    int m_staffOffset;
+    double m_staffOffset;
 
     double m_currentX;
